Reorganize recipe elements and add images to links
Images are added to the link so they can also be clicked to direct users
to the recipes. Remove underlining and change colors for links for
aesthetic purposes. Make link containers into flex boxes so that images
and text within the link can be vertically aligned and centered.

Plan to:
 - Fix how images are displayed so they can be the same size but cropped
   and not distorted